Made changes to fix the game coordinates for tiles
This made it possible to get rid of the TileObstacle Class and just use
the standard Obstacle class.
Game Height was made static (not entirely, just a dirty quick fix) so
you don't have to pass the Game as an argument to all rendering classes,
but I am not sure if this is a good way to do it.
A consequence of this is that the world file data is "upside down", but
I think it is more logical, if the first row represents x = 0 and y = 0.
It also makes the Game coordinates much cleaner. Every coordinate now
references the bottom left corner in a coordinate system with the origin
in the bottom left of the world. The transformations happen ONLY in
render methods